melanoma_executor/requirements.txt
flask==1.1.2
gunicorn==20.0.4
numpy==1.19.1
tensorflow==2.3.0
pillow==7.2.0
pyheif==0.5.1